IO: add dodgy Wavefront OBJ file parser
Based on the File class, it loads all the file in memory,
constructs arrays of data and generate buffer ready to be pushed
to OpenGL or any 3D backend.

This parsing is very simple, it doesn't take in account multiple
objects, colors, materials, or wathever.
You should probably never use this code in a production
environment.