System: add StackAllocator implementation
The StackAllocator implementation is a linked list of memory pages
containing a stack of allocated slots. During a new allocation,
the top slot (free) is returned. Pages can be expanded by
allocating a new one. During deallocation, slots are unstacked and
when a page is completely unused, it's destroyed.

There is always at least one page available to avoid a malloc on
the first alloc call.

Warning: not thread-safe!

This is an experimental code written years ago, it has not been
retested.